requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
openai==0.28.0
pydantic==2.5.2
python-multipart==0.0.6
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.1.0
asyncio==3.4.3
httpx==0.25.2
lxml==4.9.3